[feaure/PPF-10]: rework entities
