+++ src/features/Auth/model/index.ts
@@ -1,3 +1,2 @@
 // Экспорт всех частей модели аутентификации
 export * from './constants';
-export * from './config';
