+++ src/shared/ui/AuthForm/AuthForm.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Form, Input, Button, Alert } from "antd";
 import { EyeInvisibleOutlined, EyeOutlined } from "@ant-design/icons";
 import type { FormFieldConfig, BaseFormProps } from "../../lib/forms";
